components/data_collector/data_extractor/requirements.txt
Pillow==10.4.0
piexif==1.1.3
imagehash==4.3.1
